feat(PE-3607): add endpoint to fetch all contract interactions
This endpiont returns all the contract interactions for a given contract, with priority given to graphql. We may want to validate the interctions are the same from graphql and warp, but for now we'll use graphql to tell us what's relevant and use warp to determine if they were successful or failed. So long as the warp on configuration and the arweave client use the same gateway, they *should* return the same interactions.