feat(PE-3788): validate contracts based on state values
This is a simple validation check for ant contracts. We may consider using something more advanaced like ajv or zod, but this is good enough for now. It additionally adds constraints for evaluation timeouts, limiting to 5 secs. It's likely most constraints will be below this constraint other than some larger ones lke the ArNS registry.