fix(errors): add error middleware that handles erorrs thrown, based on their types and returns proper status codes
This type of middleware is particularly useful for this type of problem. We have various instances in which warp throws errors when failing to evaluate a contract state, and we want to parse those properly and return the correct
status code when they occur. Right now we over thrown 503 errors, which suggest to clients that the app failed to proc
ess, when it's more likely they are requesting contracts that have not been mined yet or do not adhere to specs define
d by warp.