fix(regex): remove field regex for contract fields
We were using a regex to prevent arbitaray requests creating potential issues in the service. Given the expanding set of fields, and the desire to support ant contracts that may have extended scheams, we can remove this regex in the service. If the field does not exist on the state, a 404 will be returned.